Simplify how old init script is backed up.
I have been getting errors when running repeated highstates, where
salt tries to "back up" the old script from /etc/init.d/nginx to
/usr/share/nginx/init.d . Since the symlink was added, the source
file still exists, but the rename fails because the destination
already exists. This causes a series of cascading errors in states
that depend on this on the file rename state.